+++ src/components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import { TextLink } from '@/components/TextLink/TextLink';
-import { currentYear } from '@/utils/dates';
-import { config } from '@/utils/config';
+import { currentYear } from '@/lib/dates';
+import { config } from '@/lib/config';
 
 export const Footer = () => (
     <footer className="container mx-auto mb-3 mt-8 flex max-w-screen-md justify-between border-t p-3 text-sm text-gray-400">
