+++ src/features/account/hooks/useForgotPassword.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { ForgotPasswordRequest } from '@/features/account/accountTypes';
-import { fetcher } from '@/utils/fetch';
-import { config } from '@/utils/config';
+import { fetcher } from '@/lib/fetch';
+import { config } from '@/lib/config';
 
 export const forgotPassword = (request: ForgotPasswordRequest) =>
     fetcher(`${config.API_URL}/account/forgot-password`, {
