+++ src/features/account/hooks/useRegister.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { UpdatedResponse } from '@/features/common/commonTypes';
 import { RegisterRequest } from '@/features/account/accountTypes';
-import { fetcher } from '@/utils/fetch';
-import { config } from '@/utils/config';
+import { fetcher } from '@/lib/fetch';
+import { config } from '@/lib/config';
 
 const register = (request: RegisterRequest) =>
     fetcher<UpdatedResponse>(`${config.API_URL}/account/register`, {
