+++ src/features/manage/hooks/useChangePassword.ts
@@ -2,8 +2,8 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useSession } from 'next-auth/react';
 import { UpdatedResponse } from '@/features/common/commonTypes';
 import { ChangePasswordRequest } from '@/features/manage/manageTypes';
-import { fetcher } from '@/utils/fetch';
-import { config } from '@/utils/config';
+import { fetcher } from '@/lib/fetch';
+import { config } from '@/lib/config';
 
 const changePassword = (request: ChangePasswordRequest, init?: RequestInit) =>
     fetcher<UpdatedResponse>(`${config.API_URL}/manage/change-password`, {
