+++ src/features/manage/hooks/useUpdateProfile.ts
@@ -2,8 +2,8 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useSession } from 'next-auth/react';
 import { UpdatedResponse } from '@/features/common/commonTypes';
 import { UpdateProfileRequest } from '@/features/manage/manageTypes';
-import { fetcher } from '@/utils/fetch';
-import { config } from '@/utils/config';
+import { fetcher } from '@/lib/fetch';
+import { config } from '@/lib/config';
 
 const updateProfile = (request: UpdateProfileRequest, init?: RequestInit) =>
     fetcher<UpdatedResponse>(`${config.API_URL}/manage`, {
