+++ src/features/user/CreateUserForm/CreateUserForm.tsx
@@ -6,8 +6,8 @@ import { DatePicker } from '@/components/Form/DatePicker';
 import { ToggleGroup } from '@/components/Form/ToggleGroup';
 import { Button } from '@/components/Button/Button';
 import { ButtonGroup } from '@/components/Button/ButtonGroup';
-import { Role, roleOptions } from '@/utils/auth';
-import { isInPast } from '@/utils/dates';
+import { Role, roleOptions } from '@/lib/auth';
+import { isInPast } from '@/lib/dates';
 
 const schema = z
     .object({
