+++ src/features/user/hooks/useGetUser.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { useSession } from 'next-auth/react';
 import { GetUserResponse } from '@/features/user/userTypes';
-import { fetcher } from '@/utils/fetch';
-import { config } from '@/utils/config';
+import { fetcher } from '@/lib/fetch';
+import { config } from '@/lib/config';
 
 export const getUser = (id: string, init?: RequestInit) =>
     fetcher<GetUserResponse>(`${config.API_URL}/user/${id}`, init);
